clang-tidy changes

+++ onnxruntime/core/providers/cpu/tensor/compress.cc
@@ -74,13 +74,13 @@ Status Compress::Compute(OpKernelContext* ctx) const {
     if (!IAllocator::CalcMemSizeForArray(static_cast<size_t>(axes_right_stride), element_bytes,
                                          &axes_right_stride_bytes))
       return Status(ONNXRUNTIME, FAIL, "size overflow");
-    for (int i = 0; i < axes_left_stride; ++i) {
-      for (int j = 0; j < valid_condition_length; ++j) {
+    for (int64_t i = 0; i < axes_left_stride; ++i) {
+      for (int64_t j = 0; j < valid_condition_length; ++j) {
         if (!condition_data[j]) {
           continue;
         }
         if (is_string_type) {
-          for (int idxItem = 0; idxItem < axes_right_stride; ++idxItem) {
+          for (int64_t idxItem = 0; idxItem < axes_right_stride; ++idxItem) {
             reinterpret_cast<std::string*>(output_data)[output_index + idxItem] =
               reinterpret_cast<const std::string*>(input_data)[i * axes_included_right_stride + j * axes_right_stride + idxItem];
           }
@@ -92,7 +92,7 @@ Status Compress::Compute(OpKernelContext* ctx) const {
       }
     }
   } else {
-    for (int i = 0; i < valid_condition_length; ++i) {
+    for (int64_t i = 0; i < valid_condition_length; ++i) {
       if (!condition_data[i]) {
         continue;
       }
